Pre merge branch (#5)
move methods utils , use async load of helper, review comments
