+++ libs/utils/utils.js
@@ -1,9 +1,5 @@
 /* eslint-disable no-console */
 
-import { enablePersonalizationV2, loadAnalyticsAndInteractionData, getEnv, getMetadata } from '../martech/helpers.js';
-
-export { getMetadata };
-
 const MILO_TEMPLATES = [
   '404',
   'featured-story',
@@ -118,6 +114,31 @@ const DO_NOT_INLINE = [
   'z-pattern',
 ];
 
+const ENVS = {
+  stage: {
+    name: 'stage',
+    ims: 'stg1',
+    adobeIO: 'cc-collab-stage.adobe.io',
+    adminconsole: 'stage.adminconsole.adobe.com',
+    account: 'stage.account.adobe.com',
+    edgeConfigId: '8d2805dd-85bf-4748-82eb-f99fdad117a6',
+    pdfViewerClientId: '600a4521c23d4c7eb9c7b039bee534a0',
+  },
+  prod: {
+    name: 'prod',
+    ims: 'prod',
+    adobeIO: 'cc-collab.adobe.io',
+    adminconsole: 'adminconsole.adobe.com',
+    account: 'account.adobe.com',
+    edgeConfigId: '2cba807b-7430-41ae-9aac-db2b0da742d5',
+    pdfViewerClientId: '3c0a5ddf2cc04d3198d9e48efc390fa9',
+  },
+};
+ENVS.local = {
+  ...ENVS.stage,
+  name: 'local',
+};
+
 export const MILO_EVENTS = { DEFERRED: 'milo:deferred' };
 
 const LANGSTORE = 'langstore';
@@ -126,6 +147,29 @@ const PAGE_URL = new URL(window.location.href);
 export const SLD = PAGE_URL.hostname.includes('.aem.') ? 'aem' : 'hlx';
 
 const PROMO_PARAM = 'promo';
+
+export function getEnv(conf) {
+  const { host } = window.location;
+  const query = PAGE_URL.searchParams.get('env');
+
+  if (query) return { ...ENVS[query], consumer: conf[query] };
+
+  const { clientEnv } = conf;
+  if (clientEnv) return { ...ENVS[clientEnv], consumer: conf[clientEnv] };
+
+  if (host.includes('localhost')) return { ...ENVS.local, consumer: conf.local };
+  /* c8 ignore start */
+  if (host.includes(`${SLD}.page`)
+    || host.includes(`${SLD}.live`)
+    || host.includes('stage.adobe')
+    || host.includes('corp.adobe')
+    || host.includes('graybox.adobe')) {
+    return { ...ENVS.stage, consumer: conf.stage };
+  }
+  return { ...ENVS.prod, consumer: conf.prod };
+  /* c8 ignore stop */
+}
+
 let delayedMartech = false;
 
 export function getLocale(locales, pathname = window.location.pathname) {
@@ -145,6 +189,12 @@ export function getLocale(locales, pathname = window.location.pathname) {
   locale.prefix = isUS ? '' : `/${localeString}`;
   locale.region = isUS ? 'us' : localeString.split('_')[0];
   return locale;
+}
+
+export function getMetadata(name, doc = document) {
+  const attr = name && name.includes(':') ? 'property' : 'name';
+  const meta = doc.head.querySelector(`meta[${attr}="${name}"]`);
+  return meta && meta.content;
 }
 
 const handleEntitlements = (() => {
@@ -983,6 +1033,38 @@ export async function loadMartech({
   return true;
 }
 
+/**
+ * Checks if the user is signed out based on the server timing and navigation performance.
+ * 
+ * @returns {boolean} True if the user is signed out, otherwise false.
+ */
+function isSignedOut() {
+  let w = window, perf = w.performance, serverTiming = {};
+
+  if (perf && perf.getEntriesByType) {
+    serverTiming = Object.fromEntries(
+      perf.getEntriesByType("navigation")?.[0]?.serverTiming?.map?.(
+        ({ name, description }) => ([name, description])
+      ) ?? []
+    );
+  }
+
+  const isSignedOutOnStagingOrProd = serverTiming && serverTiming.sis === '0';
+
+  // Return true if it's a dev environment or signed out on staging/prod
+  return !Object.keys(serverTiming || {}).length || isSignedOutOnStagingOrProd;
+}
+
+/**
+ * Enables personalization (V2) for the page.
+ * 
+ * @returns {boolean} True if personalization is enabled, otherwise false.
+ */
+export function enablePersonalizationV2() {
+  const enablePersV2 = document.head.querySelector(`meta[name="personalization-v2"]`);
+  return enablePersV2 && isSignedOut();
+}
+
 async function checkForPageMods() {
   const {
     mep: mepParam,
@@ -990,21 +1072,24 @@ async function checkForPageMods() {
     mepButton,
     martech,
   } = Object.fromEntries(PAGE_URL.searchParams);
-  let targetInteractionData = null;
+  let targetInteractionPromise = null;
   if (mepParam === 'off') return;
   const pzn = getMepEnablement('personalization');
   const promo = getMepEnablement('manifestnames', PROMO_PARAM);
   const target = martech === 'off' ? false : getMepEnablement('target');
   const xlg = martech === 'off' ? false : getMepEnablement('xlg');
-  const enablePersV2 = enablePersonalizationV2();
 
   if (!(pzn || target || promo || mepParam
     || mepHighlight || mepButton || mepParam === '' || xlg)) return;
-
+    
+  const enablePersV2 = enablePersonalizationV2();
   if (martech !== 'off' && (target || xlg || pzn) && enablePersV2) {
     const { locale } = getConfig();
     try{
-      targetInteractionData = await loadAnalyticsAndInteractionData({ locale });
+      const { loadAnalyticsAndInteractionData } = await import('../martech/helpers.js');
+      targetInteractionPromise = loadAnalyticsAndInteractionData(
+        { locale, env: getEnv({})?.name, timeoutMeta: getMetadata('target-timeout') }
+      );
     } catch (err){
       console.log('Interact Call didnt go through', err);
     }
@@ -1023,7 +1108,7 @@ async function checkForPageMods() {
   const { init } = await import('../features/personalization/personalization.js');
   await init({
     mepParam, mepHighlight, mepButton, pzn, promo, target,
-  }, targetInteractionData);
+  }, targetInteractionPromise);
 }
 
 async function loadPostLCP(config) {
